moved fen tests into parse.rs
